chore(qtimmodule): use text input v3
* add keyboard
* add text input v3
* delete dim text input v1
