+++ src/qtimmodule/qt5/Keyboard.cpp
@@ -1,0 +1,1 @@
+../qt6/Keyboard.cpp
