+++ src/qtimmodule/qt5/Keyboard.h
@@ -1,0 +1,1 @@
+../qt6/Keyboard.h
