+++ src/qtimmodule/qt5/TextInputV3.h
@@ -1,0 +1,1 @@
+../qt6/TextInputV3.h
