+++ src/qtimmodule/qt5/plugin.json
@@ -1,5 +1,1 @@
-{
-    "Keys": [
-        "dim"
-    ]
-}
+../qt6/plugin.json
