--- src/qtimmodule/qt6/DimTextInputV1.cpp
@@ -1,1 +1,0 @@
-../qt5/DimTextInputV1.cpp
