--- src/qtimmodule/qt6/DimTextInputV1.h
@@ -1,1 +1,0 @@
-../qt5/DimTextInputV1.h
