--- src/qtimmodule/qt6/Key.cpp
@@ -1,1 +1,0 @@
-../qt5/Key.cpp
