--- src/qtimmodule/qt6/Key.h
@@ -1,1 +1,0 @@
-../qt5/Key.h
