+++ src/qtimmodule/qt6/plugin.json
@@ -1,1 +1,5 @@
-../qt5/plugin.json
+{
+    "Keys": [
+        "dim"
+    ]
+}
