+++ src/wl/client/ZwpTextInputManagerV3.cpp
@@ -19,9 +19,7 @@ ZwpTextInputManagerV3::~ZwpTextInputManagerV3()
     zwp_text_input_manager_v3_destroy(get());
 }
 
-std::shared_ptr<ZwpTextInputV3>
-ZwpTextInputManagerV3::get_text_input(const std::shared_ptr<Seat> &seat)
+struct zwp_text_input_v3 *ZwpTextInputManagerV3::get_text_input(const std::shared_ptr<Seat> &seat)
 {
-    return std::make_shared<ZwpTextInputV3>(
-        zwp_text_input_manager_v3_get_text_input(get(), seat->get()));
+    return zwp_text_input_manager_v3_get_text_input(get(), seat->get());
 }
