+++ src/wl/client/ZwpTextInputV3.cpp
@@ -6,9 +6,19 @@
 
 using namespace wl::client;
 
+const zwp_text_input_v3_listener ZwpTextInputV3::listener_ = {
+    ListenerCallbackWrapper<&ZwpTextInputV3::zwp_text_input_v3_enter>::func,
+    ListenerCallbackWrapper<&ZwpTextInputV3::zwp_text_input_v3_leave>::func,
+    ListenerCallbackWrapper<&ZwpTextInputV3::zwp_text_input_v3_preedit_string>::func,
+    ListenerCallbackWrapper<&ZwpTextInputV3::zwp_text_input_v3_commit_string>::func,
+    ListenerCallbackWrapper<&ZwpTextInputV3::zwp_text_input_v3_delete_surrounding_text>::func,
+    ListenerCallbackWrapper<&ZwpTextInputV3::zwp_text_input_v3_done>::func,
+};
+
 ZwpTextInputV3::ZwpTextInputV3(zwp_text_input_v3 *val)
     : Type(val)
 {
+    zwp_text_input_v3_add_listener(get(), &listener_, this);
 }
 
 ZwpTextInputV3::~ZwpTextInputV3()
